+++ src/main/java/rldevs4j/agents/ppo/PPOActor.java
@@ -8,6 +8,6 @@ public interface PPOActor {
     public void saveModel(String path) throws IOException;
     public void loadModel(String path) throws IOException;
     public INDArray[] output(INDArray obs, INDArray act);
-    public double[] action(INDArray obs);
+    public float[] action(INDArray obs);
     public double train(INDArray states, INDArray actions, INDArray advantages, INDArray logOldPi, int iteration, int epoch);
 }
